% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_data}
\alias{country_data}
\alias{mortality}
\title{Mortality rates for 47 countries or sub-populations, various years}
\format{A tibble with 47 rows and 3 columns
\describe{
\item{\code{country}}{Country name. This includes sub-units such as
    the NZ Maori population.}
\item{\code{ccode}}{Country or unit code}
\item{\code{data}}{A list column containing the data for that
    country. Consists of five columns: year, age, female, male, and
    total, with the latter three being estimated mortality rates
    for the year and age combination.}
}}
\source{
mortality.org
}
\usage{
mortality
}
\description{
Mortality rates for 47 countries or sub-populations, various years
}
\keyword{datasets}
