% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ireland}
\alias{ireland}
\title{Mortality rates in Ireland 1950-2014}
\format{A tibble with 7,215 rows and 5 columns
\describe{
\item{\code{year}}{Year}
\item{\code{age}}{Age}
\item{\code{female}}{Female mortality rate}
\item{\code{male}}{Male mortality rate}
\item{\code{total}}{Total morality rate}
}}
\source{
mortality.org
}
\usage{
ireland
}
\description{
Mortality rates in Ireland 1950-2014
}
\keyword{datasets}
