% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sweden}
\alias{sweden}
\title{Mortality rates in Sweden}
\format{A tibble
\describe{
\item{\code{year}}{Year}
\item{\code{age}}{Age}
\item{\code{female}}{Female mortality rate}
\item{\code{male}}{Male mortality rate}
\item{\code{total}}{Total morality rate}
}}
\source{
mortality.org
}
\usage{
sweden
}
\description{
Mortality rates in Sweden
}
\keyword{datasets}
