% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{britain}
\alias{britain}
\title{Mortality rates in Britain}
\format{A tibble with 19,536 rows and 5 columns
\describe{
\item{\code{year}}{Year}
\item{\code{age}}{Age}
\item{\code{female}}{Female mortality rate}
\item{\code{male}}{Male mortality rate}
\item{\code{total}}{Total morality rate}
}}
\source{
mortality.org
}
\usage{
britain
}
\description{
Mortality rates in Britain 1841-2016
}
\keyword{datasets}
