% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{france}
\alias{france}
\title{Mortality rates in France 1816-2016}
\format{A tibble with 22,311 rows and 5 columns
\describe{
\item{\code{year}}{Year}
\item{\code{age}}{Age}
\item{\code{female}}{Female mortality rate}
\item{\code{male}}{Male mortality rate}
\item{\code{total}}{Total morality rate}
}}
\source{
mortality.org
}
\usage{
france
}
\description{
Mortality rates in France 1816-2016
}
\keyword{datasets}
