% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nzl}
\alias{nzl}
\title{Mortality rates in New Zealand}
\format{A tibble
\describe{
\item{\code{year}}{Year}
\item{\code{age}}{Age}
\item{\code{female}}{Female mortality rate}
\item{\code{male}}{Male mortality rate}
\item{\code{total}}{Total morality rate}
}}
\source{
mortality.org
}
\usage{
nzl
}
\description{
Mortality rates in New Zealand
}
\keyword{datasets}
