% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{okboomer}
\alias{okboomer}
\title{Birth rates in the US and England/Wales 1938-2015}
\format{A tibble
\describe{
\item{\code{year}}{Year}
\item{\code{month}}{Month}
\item{\code{n_days}}{N days in month}
\item{\code{births}}{N births}
\item{\code{total_pop}}{National Population}
\item{\code{births_pct}}{N births / National Population}
\item{\code{births_pct_day}}{Average number of births per million
    people per day, for the observed month}
\item{\code{date}}{Date}
}}
\source{
US Census Bureau and UK ONS
}
\usage{
okboomer
}
\description{
Birth rates in the US and England/Wales 1938-2015
}
\keyword{datasets}
